♻️ Move Pf2eWeaponData to Json2QuteItem
